+++ test/cookbooks/test/recipes/yum_repo.rb
@@ -1,10 +1,7 @@
 # Before that, we use "native" versions
 
-unless node['mysql_test'].nil?
-  case node['mysql_test']['version']
-  when '5.6'
-    include_recipe 'yum-mysql-community::mysql56'
-  when '5.7'
-    include_recipe 'yum-mysql-community::mysql57'
+unless node['mysql_test'].nil? || platform_family?('debian')
+  yum_mysql_community_repo 'mysql_repo' do
+    version node['mysql_test']['version']
   end
 end